refactor(ast-types): extract the shared AST definitions into a package
The linter, minifier and prettier plugin each carried their own copy of
the node definitions for the JSON tree the wasm parser emits, and the
three copies had already drifted. A fourth consumer (the type checker)
would have made that four.

`@gengjiawen/monkey-ast-types` becomes the single source of truth. The
consumers keep their `types.ts` as a re-export plus whatever is genuinely
local to them, so no import site changes. The package also carries the
type-annotation nodes the parser gains next, along with
`isTypeAnnotation` and `printTypeAnnotation` so tools can skip or render
those subtrees without reimplementing the grammar.

One behavioural note: `raw` on an integer literal is typed for both wasm
entries, since `analyze_lossless` keeps the source text while `parse`
has already produced a number. Call sites that assumed a string now say
so with `String(...)`.
